chore: Move getChatMessages API to ChatAuthApi
